+++ internal/application/error.go
@@ -3,7 +3,7 @@ package application
 import "fmt"
 
 var (
-	errNotChatMember = func(chatID, userID int32) error {
+	errNotChatMember = func(chatID, userID int64) error {
 		return fmt.Errorf("user %d is not a member of chat %d", userID, chatID)
 	}
 )
